# standard/backend/docs/swagger.yaml
basePath: /api/v1/standard
definitions:
  client.DiscoverableAsset:
    properties:
      description:
        description: 描述（可空）
        type: string
      name:
        description: 资源名称
        type: string
      source_reference:
        description: 在源模块中的唯一标识
        type: string
    type: object
...
  github_com_addp_standard_internal_models.UpdateDomainRequest:
    properties:
      description:
        type: string
      icon:
        type: string
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      version:
        type: integer
    required:
    - version
    type: object
...
  github_com_addp_standard_internal_service.DomainTree:
    properties:
      children:
        items:
          $ref: '#/definitions/github_com_addp_standard_internal_service.DomainTree'
        type: array
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
      version:
        type: integer
    type: object
host: localhost:8110
info:
  contact: {}
  title: ADDP Standard API
  version: "1.0"
paths:
  /assets/discoverable:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/client.DiscoverableAsset'
            type: array
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 可发现资产列表 | List discoverable assets
      tags:
      - AssetDiscoverable
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.metric.read
...
  /domains:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_addp_standard_internal_service.DomainTree'
            type: array
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取业务域列表 | List business domains
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.domain.read
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建业务域 | Create business domain
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.domain.create
  /domains/{id}:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除业务域 | Delete business domain
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.domain.delete
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取业务域详情 | Get business domain detail
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.domain.read
    put:
      parameters:
      - description: 更新业务域 | Update business domain
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_addp_standard_internal_models.UpdateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 资源版本冲突 | Resource version conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新业务域 | Update business domain
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.domain.update
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"